# lib/api-zod/src/generated/types/paper.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Paper {
  id: number;
  examName: string;
  year?: string | null;
  shift?: string | null;
  totalQuestions: number;
  fileName?: string | null;
  createdAt?: string;
}