# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  paperId?: number | null;
  questionNumber: number;
  questionIdOriginal?: string | null;
  questionText: string;
  optionA?: string | null;
  optionB?: string | null;
  optionC?: string | null;
  optionD?: string | null;
  correctAnswer?: string | null;
  chosenOption?: string | null;
  status?: string | null;
  hasFigure?: boolean;
  figureData?: string | null;
  subject?: string | null;
  note?: string | null;
  createdAt?: string;
}