# lib/api-zod/src/generated/types/questionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionStatsBySubjectItem } from "./questionStatsBySubjectItem";

export interface QuestionStats {
  totalPapers: number;
  totalQuestions: number;
  withFigures: number;
  bySubject: QuestionStatsBySubjectItem[];
}